+++ .github/workflows/ci.yml
@@ -3,7 +3,7 @@ name: Turners AutoMatchmaker Build & Test
 on:
   push:
     branches:
-      - main
+      - development/Kelsie
 
 env:
   AZURE_WEBAPP_NAME: "M3-Prediction-API"
